# src/swagger.yaml
openapi: 3.0.3
info:
  title: Market Data Sync and Stats API
  version: 1.1.0
  description: >
    APIs for syncing 52-week stats, daily & 52-week market breadth, daily data, and fetching instrument stats.
servers:
  - url: http://localhost:3015
    description: Local dev server
  - url: https://trade-dashboard-backend-y4jp.onrender.com
    description: QA server
paths:
...
  /upstoxs/initiate-token:
    get:
      summary: Initiate Upstox access token request
      description: Triggers the process to initiate an access token request to Upstox.
      responses:
        '200':
          description: Token initiation triggered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '500':
          description: Server error
...